requirements_api.txt
# Jalikoi Analytics - Required Python Packages
# Install with: pip install -r requirements.txt

# Core Data Analysis
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Excel Export
openpyxl>=3.0.0

# Database Support (Optional - for MySQL connection)
mysql-connector-python>=8.0.0
pymysql>=1.0.0
sqlalchemy>=1.4.0

# REST API Support (for FastAPI)
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
